review detail UI / refactor
