+++ app/src/main/java/com/example/wishhair/review/detail/ReviewDetailActivity.java
@@ -1,0 +1,16 @@
+package com.example.wishhair.review.detail;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+import com.example.wishhair.R;
+
+public class ReviewDetailActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_review_detail);
+    }
+}
