+++ app/src/main/java/com/example/wishhair/review/recent/RecyclerViewAdapterRecent.java
@@ -1,4 +1,4 @@
-package com.example.wishhair.review;
+package com.example.wishhair.review.recent;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -12,6 +12,7 @@ import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.wishhair.R;
+import com.example.wishhair.review.ReviewItem;
 
 import java.util.ArrayList;
