+++ app/src/main/res/layout/activity_review_detail.xml
@@ -1,0 +1,97 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".review.detail.ReviewDetailActivity"
+    android:orientation="vertical">
+
+<!--    image-->
+    <androidx.viewpager2.widget.ViewPager2
+        android:id="@+id/review_detail_viewPager"
+        android:layout_width="match_parent"
+        android:layout_height="300dp" />
+    <LinearLayout
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:orientation="vertical"
+        android:padding="20dp">
+        <!--    title -->
+        <TextView
+            android:id="@+id/review_detail_hairStyle"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="헤어스타일 이름"
+            android:textSize="30sp"
+            android:textStyle="bold"/>
+        <LinearLayout
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content">
+            <TextView
+                android:id="@+id/review_detail_tags"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="#태그들"
+                android:textStyle="bold"/>
+        </LinearLayout>
+        <!--    grade, heart -->
+        <LinearLayout
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="10dp">
+            <ImageView
+                android:id="@+id/review_detail_imageView_gradeMark"
+                android:layout_width="20dp"
+                android:layout_height="20dp"
+                android:layout_marginEnd="5dp"
+                android:background="@drawable/star_fill"
+                tools:ignore="ContentDescription" />
+
+            <TextView
+                android:id="@+id/review_detail_tv_grade"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="3.04" />
+
+            <ImageView
+                android:id="@+id/review_detail_imageView_Heart"
+                android:layout_width="20dp"
+                android:layout_height="20dp"
+                android:layout_marginStart="25dp"
+                android:layout_marginEnd="5dp"
+                android:src="@drawable/heart_fill"
+                tools:ignore="ContentDescription" />
+
+            <TextView
+                android:id="@+id/review_detail_tv_heartCount"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="500" />
+
+        </LinearLayout>
+        <!--    date -->
+        <LinearLayout
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="10dp">
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="작성일 "/>
+
+            <TextView
+                android:id="@+id/review_detail_tv_date"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="yyyy. mm. dd" />
+        </LinearLayout>
+<!--        content -->
+        <TextView
+            android:id="@+id/review_detail_tv_content"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="10dp"
+            android:text="내용"/>
+    </LinearLayout>
+
+</LinearLayout>
