+++ app/src/main/res/layout/fragment_review_list.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".review.ReviewListFragment">
+    tools:context=".review.recent.ReviewListFragment">
 
     <LinearLayout
         android:id="@+id/review_parentLinearLayout"
