+++ app/src/main/res/layout/fragment_review_my.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".review.ReviewMyFragment">
+    tools:context=".review.my.ReviewMyFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/review_my_recyclerView"
